Add group user search and invite functionality
A new feature has been added that allows the search of group users and sending of invites to join an event. Changes involve the creation of a new component, EventInvites, and updates to API service and event page. Additionally, a minor bug fix was done to correct an API endpoint in updateGroupUser function.